feat: 화면 렌더링 optimization
